refactor: refactored choere model specific functions to a separate folder
